TodoList component created and separated
